+++ backend/src/game/game.module.ts
@@ -1,8 +1,12 @@
 import { Module } from '@nestjs/common';
-import { GameService } from './game.service';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { GameGateway } from './game.gateway';
+import { GameService } from './game.service';
+import { RoomEntity } from '../database/entities/room.entity';
+import { PlayerEntity } from '../database/entities/player.entity';
 
 @Module({
-  providers: [GameService, GameGateway]
+  imports: [TypeOrmModule.forFeature([RoomEntity, PlayerEntity])],
+  providers: [GameGateway, GameService],
 })
 export class GameModule {}